Import DSM 7.0-41890 armada37xx linux-4.4.x.txz Base on Linux Kernal v4.4.180
tar url: https://global.synologydownload.com/download/ToolChain/Synology%20NAS%20GPL%20Source/7.0-41890/armada37xx/linux-4.4.x.txz
tar md5sum: f79749d6cd428696117d12300be1642b
from: https://archive.synology.com/download/ToolChain/Synology%20NAS%20GPL%20Source/7.0-41890

+++ drivers/usb/serial/ch341.c
@@ -68,6 +68,7 @@
 #define CH341_NBREAK_BITS_REG1 0x01
 #define CH341_NBREAK_BITS_REG2 0x40
 
+
 static const struct usb_device_id id_table[] = {
 	{ USB_DEVICE(0x4348, 0x5523) },
 	{ USB_DEVICE(0x1a86, 0x7523) },
@@ -117,7 +118,7 @@ static int ch341_control_in(struct usb_device *dev,
 	r = usb_control_msg(dev, usb_rcvctrlpipe(dev, 0), request,
 			    USB_TYPE_VENDOR | USB_RECIP_DEVICE | USB_DIR_IN,
 			    value, index, buf, bufsize, DEFAULT_TIMEOUT);
-	if (r < bufsize) {
+	if (r < (int)bufsize) {
 		if (r >= 0) {
 			dev_err(&dev->dev,
 				"short control message received (%d < %u)\n",
@@ -312,6 +313,7 @@ static void ch341_close(struct usb_serial_port *port)
 	usb_serial_generic_close(port);
 	usb_kill_urb(port->interrupt_in_urb);
 }
+
 
 /* open this device, set default parameters */
 static int ch341_open(struct tty_struct *tty, struct usb_serial_port *port)
